+++ src/config.py
@@ -1,0 +1,5 @@
+class cfg:
+    debug = True
+    FPS = 120
+    SCREEN_WIDTH = 400
+    SCREEN_HEIGHT = 600
